test: enable WorkingSetView-integ-test
