+++ test/UnitTestSuite.js
@@ -92,7 +92,7 @@ define(function (require, exports, module) {
     require("spec/ProjectManager-integ-test");
     require("spec/QuickOpen-integ-test");
     require("spec/ViewCommandHandlers-integ-test");
-    // require("spec/WorkingSetView-test");
+    require("spec/WorkingSetView-integ-test");
     // require("spec/WorkingSetSort-test");
     // require("spec/FindInFiles-test");
     // todo TEST_MODERN
